# 3874 use several cells to index