* The index on the `k` cell is generalised to an index on one or more arbitrary cells containing K sequences (default stays the same, the `k` cell only). 
* The cell(s) to index are read from a definition attribute named `indexCells`. This attribute is expected to contain an argument list of names _quoted as all symbols in kore_  (e.g., `Lbl'-LT-'k'-GT-'` for the `k` cell).
* Alternatively, the cell(s) to index can be overridden by a command line option `--index-cells CELL1,CELL2,...` to `booster-dev` and `kore-rpc-booster`.  This option takes plain names (like `k` for the `k` cell), which are quoted and prefixed before overriding the definition attribute of the loaded definition.

Fixes #3874