@Baltoli 2024-05-23
When this is merged we can look at getting support for the `indexCells` attribute added to the frontend. I suspect what we'll want to do is:
- Allow `index` as a K cell attribute
- Add a pass before generating KORE that pulls `index` off each configuration cell and composes the correctly-formatted `indexCells` attribute for the definition.