@jberthold 2024-05-23
| Test                                   | 3874-use-several-cells-to-index time | master-e62fba2f0 time | (3874-use-several-cells-to-index/master-e62fba2f0) time 
|----------------------------------------|--------------------------------------|-----------------------|---------------------------------------------------------
| erc20/hkg/totalSupply-spec.k           | 35.34                                | 43.83                 | 0.806297056810404                                       
| mcd/vat-addui-fail-rough-spec.k        | 56.67                                | 69.74                 | 0.8125896185833095                                      
| erc20/hkg/transfer-failure-1-spec.k    | 45.77                                | 54.29                 | 0.8430650211825383                                      
| mcd/end-cash-pass-rough-spec.k         | 186.62                               | 214.96                | 0.868161518422032                                       
| erc20/hkg/transfer-failure-2-spec.k    | 45.66                                | 52.56                 | 0.8687214611872145                                      
| erc20/hkg/transfer-success-1-spec.k    | 41.13                                | 46.83                 | 0.8782831518257528                                      
| mcd/vat-addui-pass-spec.k              | 93.52                                | 103.06                | 0.9074325635552105                                      
| mcd/end-pack-pass-rough-spec.k         | 129.96                               | 141.78                | 0.9166314007617435                                      
| erc20/hkg/balanceOf-spec.k             | 36.07                                | 39.26                 | 0.9187468160978095                                      
| mcd/pot-join-pass-rough-spec.k         | 166.61                               | 177.79                | 0.9371168232184038                                      
| erc20/hkg/transfer-success-2-spec.k    | 40.39                                | 41.94                 | 0.9630424415832142                                      
| erc20/ds/transferFrom-success-1-spec.k | 65.58                                | 63.29                 | 1.0361826512877232                                      
| mcd/vat-subui-pass-rough-spec.k        | 96.87                                | 90.61                 | 1.0690872972078138                                      
| TOTAL                                  | 1040.19                              | 1139.94    | 0.9124953944944473