@jberthold 2024-05-23
> Can you post information about how much this improves the mx-semantics repository verification capabilities?

I ran some tests locally using  `--index-cells instrs,commands,k`:

| Test                          | index `instrs,commands,k` | default `k` |
|-------------------------------|----------------|--------|
| test_add_liquidity-spec.json  |          13965 |  15515 |
| test_change_quorum-spec.json  |           2206 |   2764 |
| test_call_add-spec.json       |            126 |    147 |
| test_call_add_twice-spec.json |            204 |    242 |
| test_fund-spec.json           |            188 |    220 |

These runs were from an `mx-backend` branch that got merged today. 
The `add_liquidity` "default" run was from yesterday, without latest `mx-backend` tweaks.