@ehildenb 2024-05-23
Awesome, thanks. Have you tried different orders like `commands,instrs,k`, or `commands,k,instrs`?