@jberthold 2024-05-24
> Awesome, thanks. Have you tried different orders like `commands,instrs,k`, or `commands,k,instrs`?

I chose this order because `instrs` is the most discriminating cell (leads to the least rules with index `Anything` when used in isolation). However, the order actually does not matter for that (all cell indexes are used equally when selecting the rules for a term).