@jberthold 2024-05-24
These are the `kontrol` test timings, basically just noise IMHO (parallelism = 3)

| Test                                                                          | 3874-use-several-cells-to-index time | master-e62fba2f0 time | (3874-use-several-cells-to-index/master-e62fba2f0) time 
|-------------------------------------------------------------------------------|--------------------------------------|-----------------------|---------------------------------------------------------
| EmitContractTest.testExpectEmitDoNotCheckData()                               | 65.13                                | 75.15                 | 0.8666666666666666                                      
| ExpectRevertTest.test_expectRevert_true()                                     | 64.07                                | 68.81                 | 0.931114663566342                                       
| ExpectRevertTest.testFail_expectRevert_empty()                                | 28.76                                | 30.07                 | 0.9564349850349185                                      
| ExpectRevertTest.test_expectRevert_returnValue()                              | 115.5                                | 111.57                | 1.0352245227211616                                      
| MockCallTestFoundry.testMockCallEmptyAccount()                                | 59.09                                | 57.05                 | 1.0357581069237511                                      
| ExpectRevertTest.test_expectRevert_bytes4()                                   | 68.52                                | 66.1                  | 1.0366111951588504                                      
| PlainPrankTest.test_prank_zeroAddress_true()                                  | 101.17                               | 97.57                 | 1.0368965870656965                                      
| FeeTest.test_fee_setup()                                                      | 43.47                                | 41.61                 | 1.0447007930785868                                      
| ExpectRevertTest.test_expectRevert_message()                                  | 71.44                                | 68.26                 | 1.0465865807207735                                      
| StartPrankTestMsgSender.test_startprank_msgsender_setup()                     | 74.84                                | 71.41                 | 1.0480324884469963                                      
| test_foundry_prove.py::test_foundry_merge_nodes | 30.33                                | 28.93                 | 1.0483926719668164                                      
| ExpectCallTest.testExpectRegularCall()                                        | 66.94                                | 63.68                 | 1.0511934673366834                                      
| ExpectCallTest.testExpectStaticCall()                                         | 52.61                                | 49.56                 | 1.0615415657788538                                      
| test_foundry_prove.py::test_foundry_xml_report  | 17.79                                | 16.53                 | 1.0762250453720508                                      
| EmitContractTest.testExpectEmitLessTopics()                                   | 72.39                                | 65.83                 | 1.099650615221024                                       
| ExpectRevertTest.testFail_expectRevert_bytes4()                               | 72.51                                | 65.68                 | 1.1039890377588306                                      
| TOTAL                                                                         | 1004.56                              | 977.81     | 1.027357053006208